Validation for page

+++ public/assignment/views/page/edit-page.view.controller.js
@@ -1,9 +1,9 @@
-(function(){
+(function () {
     angular
         .module('WebAppMaker')
         .controller('EditPageController', EditPageController);
 
-    function EditPageController($location, $routeParams, PageService) {
+    function EditPageController($location, $routeParams, $window, PageService) {
         var vm = this;
         vm.userId = $routeParams.userId;
         vm.websiteId = $routeParams.websiteId;
@@ -19,25 +19,35 @@
                     vm.msg = {type: 'error', text: err.body};
                 });
         }
+
         init();
 
         function updatePage() {
-            PageService.updatePage(vm.pageId, vm.page)
-                .then(function () {
-                        $location.url('/user/' + vm.userId + '/website/' + vm.websiteId + '/page');
-                    },
-                    function (err) {
-                        vm.msg = {type: 'error', text: err.body};
-                    });
+            if (!vm.page.name) {
+                vm.msg = {type: 'error', text: 'Page name required'};
+            }
+            else {
+                PageService.updatePage(vm.pageId, vm.page)
+                    .then(function () {
+                            $location.url('/user/' + vm.userId + '/website/' + vm.websiteId + '/page');
+                        },
+                        function (err) {
+                            vm.msg = {type: 'error', text: err.body};
+                        });
+            }
         }
+
         function deletePage(pageId) {
-            var result = PageService.deletePage(pageId)
-                .then(function () {
-                        $location.url('/user/' + vm.userId + '/website/' + vm.websiteId + '/page');
-                    },
-                    function (err) {
-                        vm.msg = {type: 'error', text: err.body};
-                    });
+            var choice = $window.confirm('Are you sure you want to delete?');
+            if (choice) {
+                PageService.deletePage(pageId)
+                    .then(function () {
+                            $location.url('/user/' + vm.userId + '/website/' + vm.websiteId + '/page');
+                        },
+                        function (err) {
+                            vm.msg = {type: 'error', text: err.body};
+                        });
+            }
         }
     }
 })();
